requirements.txt
absl-py==2.1.0
attrs==23.2.0
aws-cdk-lib==2.127.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-lambda-python-alpha==2.127.0a0
cattrs==23.2.3
constructs==10.3.0
importlib-resources==6.1.1
jsii==1.94.0
numpy==1.26.4
ortools==9.8.3296
pandas==2.2.0
protobuf==4.25.2
publication==0.0.3
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2024.1
